updated model for a double recognition score

+++ server/hitchbot-secure-api/hitchbot-secure-api/Controllers/ReturnObjects/Speech.cs
@@ -20,8 +20,8 @@ namespace hitchbot_secure_api.Controllers
 
             public int? MatchAccuracy { get; set; }
             public int? EnvironmentType { get; set; }
-            public int? RecognitionScore { get; set; }
-            public int? GoogleRecognitionScore { get; set; }
+            public double? RecognitionScore { get; set; }
+            public double? GoogleRecognitionScore { get; set; }
             public int? ResponseScore { get; set; }
 
             public int? RecognizerTypeEnum { get; set; }
